Use cryptography for ECDSA operation
The `ecdsa` package is not suitable for secure applications
(https://github.com/tlsfuzzer/python-ecdsa#security).
Even though this package is used only for verifying the U2F signature we
can utilize the `cryptography` package for ECDSA operations.
Since it already handles the DER encoding/decoding of the signatures, we
can also remove some wrapper code.

Fixes #2410
